% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_curation.R
\name{createStudyTable}
\alias{createStudyTable}
\title{Create a table of all studies currently in data.frame}
\usage{
createStudyTable(bsdb.df, includeAlso = NULL)
}
\arguments{
\item{bsdb.df}{\code{data.frame} produced by \link[bugsigdbr]{importBugSigDB}, pre-filtered as desired}

\item{includeAlso}{\code{character} with column names to additionally include in the output table (default = \code{NULL})}
}
\value{
a data.frame of basic study information. Can be wrapped in 
kable_styling(kbl(...)) to format nicely.
}
\description{
Create a table of all studies currently in data.frame
}
\examples{
full.dat <- bugsigdbr::importBugSigDB()
createStudyTable(full.dat)
}
