% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{extractBugSigs}
\alias{extractBugSigs}
\title{Extract a list of signatures from a bugsigdb.org table}
\usage{
extractBugSigs(
  dat,
  direction = c("increased", "decreased"),
  tax.level = c("kingdom", "phylum", "class", "order", "family", "genus", "species",
    "strain")
)
}
\arguments{
\item{dat}{data.frame produced by \code{\link{importBugSigDB}}}

\item{direction}{One or both of c("increased", "decreased")}

\item{tax.level}{Any subset of c("kingdom", "phylum", "class", "order", "family", "genus", "species", "strain")}
}
\value{
a list of signatures
}
\description{
Extract a list of signatures from a bugsigdb.org table
}
\examples{
dat <- importBugSigDB()
extractBugSigs(dat, tax.level = "genus")
}
