% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{extractSignatures}
\alias{extractSignatures}
\title{Create a list of signatures}
\usage{
extractSignatures(
  dat,
  tax.level = "mixed",
  exact.tax.level = TRUE,
  col = "MetaPhlAn taxon names"
)
}
\arguments{
\item{dat}{A table such as output by \code{\link{bugsigdbr::importBugSigDB}}}

\item{tax.level}{Either "mixed" or any subset of c("kingdom", "phylum", "class", "order", "family", "genus", "species", "strain"). This full vector is equivalent to "mixed".}

\item{exact.tax.level}{If TRUE, return only the exact taxonomic levels specified by tax.level. FALSE is not working.}

\item{col}{Column name containing signatures. "MetaPhlAn taxon names" for bugsigdb.org.}
}
\value{
A list of signatures, with PMIDs for list element names
}
\description{
Create a list of signatures
}
\examples{
full.dat <- bugsigdbr::importBugSigDB()
extractSignatures(full.dat, tax.level="genus", exact.tax.level=TRUE, col = "MetaPhlAn taxon names")
}
