% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{frequencySigs}
\alias{frequencySigs}
\title{Identify the most frequently recurring taxa in bugsigdb}
\usage{
frequencySigs(sigs, n = 10)
}
\arguments{
\item{sigs}{a list of signatures, such as produced by \link[bugsigdbr]{getSignatures}}

\item{n}{Number of top taxa to return (default: 10)}
}
\value{
a named table of the top \code{n} most recurrent taxa (taxon names
  as names, counts as values), sorted in decreasing order
}
\description{
Identify the most frequently recurring taxa in bugsigdb
}
\examples{
dat <- bugsigdbr::importBugSigDB()
dat.select <- bugsigdbr::getSignatures(dat, tax.level = "genus", exact.tax.level=TRUE)
dat.select <- dat.select[grep("UP", names(dat.select))] #only "UP" signatures
frequencySigs(sigs=dat.select, n = 10)
}
