% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{simulateSignatures}
\alias{simulateSignatures}
\title{Simulate a list of signatures based on a universe of taxa equal in size and length to some smaller set of signatures}
\usage{
simulateSignatures(relevant.sigs, siglengths)
}
\arguments{
\item{relevant.sigs}{A list of signatures providing a relevant background for drawing taxa}

\item{siglengths}{An integer vector, the length of which provides the number of signatures to be simulated, and the
integers of which provide the number of taxa to be simulated in each signature}
}
\value{
A list of signatures, matching \code{siglengths} in number and in
  the number of taxa per signature, drawn from \code{relevant.sigs}
}
\description{
Simulate a list of signatures based on a universe of taxa equal in size and length to some smaller set of signatures
}
\examples{
full.dat <- bugsigdbr::importBugSigDB()
my.dat <- full.dat[full.dat$Curator == "Mst Afroza Parvin", ]
relevant.dat <- full.dat[full.dat$`Body site` \%in\% my.dat$`Body site`, ]
relevant.sigs <- bugsigdbr::getSignatures(relevant.dat, tax.level = "genus")
siglengths <- sapply(my.dat, length)
simulateSignatures(relevant.sigs, siglengths)
}
