% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{subsetByCondition}
\alias{subsetByCondition}
\title{Subset a data.frame of signatures by condition of interest}
\usage{
subsetByCondition(dat, condition, condition.column = "Condition")
}
\arguments{
\item{dat}{data.frame produced by \code{\link{bugsigdbr::importBugSigDB}}}

\item{condition}{health condition or disease of interest to subset by}

\item{condition.column}{name of column of conditions in data.frame}
}
\value{
data.frame subsetted by condition
}
\description{
Subset a data.frame of signatures by condition of interest
}
\examples{
full.dat <- bugsigdbr::importBugSigDB()
obese.dat <- subsetByCondition(full.dat, condition="obesity")
}
