% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{subsetByCurator}
\alias{subsetByCurator}
\title{Subset a data.frame of signatures by curator}
\usage{
subsetByCurator(dat, curator, curator.column = "Curator")
}
\arguments{
\item{dat}{data.frame produced by \link[bugsigdbr]{importBugSigDB}}

\item{curator}{character vector of one or more curator names to subset by}

\item{curator.column}{name of column of curators in data.frame (default: "Curator")}
}
\value{
data.frame subsetted by curator
}
\description{
Subset a data.frame of signatures by curator
}
\examples{
full.dat <- bugsigdbr::importBugSigDB()
fatima.dat <- subsetByCurator(full.dat, curator="Fatima Zohra")
}
