% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_curation.R
\name{createStudyTable}
\alias{createStudyTable}
\title{Create a table of all studies currently in data.frame}
\usage{
createStudyTable(dat)
}
\arguments{
\item{dat}{data.frame produced by \link[bugsigdbr]{importBugSigDB}, subsetted as desired}
}
\value{
a data.frame of basic study information. Can be wrapped in 
kable_styling(kbl(.)) to format nicely.
}
\description{
Create a table of all studies currently in data.frame
}
\examples{
full.dat <- bugsigdbr::importBugSigDB()
createStudyTable(full.dat)
## kable_styling(kbl(createStudyTable(full.dat))) #for html styling
}
