% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_curation.R
\name{createTaxonTable}
\alias{createTaxonTable}
\title{Create a table of most frequent taxa in a data.frame}
\usage{
createTaxonTable(dat, sig.type = c("increased, decreased, both"), n = 10)
}
\arguments{
\item{dat}{data.frame produced by \link[bugsigdbr]{importBugSigDB}, subsetted as desired}

\item{sig.type}{increased for increased in cases relative to controls, decreased for decreased in cases relative to controls, both for either}

\item{n}{number of taxa to return (if sig.type=="both", this is the number of taxa to return for each direction)}
}
\value{
kable table
}
\description{
Create a table of most frequent taxa in a data.frame
}
\examples{
full.dat <- bugsigdbr::importBugSigDB()
createTaxonTable(full.dat, n=20)
}
