% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{.make_unique_study_ID}
\alias{.make_unique_study_ID}
\title{Author: Giacomo Antonello
Date: 2025-03-17}
\usage{
.make_unique_study_ID(bsdb.df)
}
\arguments{
\item{bsdb.df}{\code{data.frame} produced by \link[bugsigdbr]{importBugSigDB}, pre-filtered as desired}
}
\description{
Description:
}
\details{
This function takes a raw bugSigDB input from `bugsigdbr` and generates a 
unique idenfier as curatedMetagenomicsData does: full last name, initial(s) of 
first name(s) and year of publication. Additionally, it checks if there are 
more PMID codes associated with the same ID and adds a .1, .2, for each 
duplication
}
