% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{frequencySigs}
\alias{frequencySigs}
\title{Identify the most frequently recurring taxa in bugsigdb}
\usage{
frequencySigs(sigs, n = 10)
}
\arguments{
\item{sigs}{data.frame produced by \link[bugsigdbr]{importBugSigDB}}

\item{n}{Number of top taxa to return}
}
\value{
integer vector of top most recurrent taxa
}
\description{
Identify the most frequently recurring taxa in bugsigdb
}
\examples{
dat <- bugsigdbr::importBugSigDB()
dat.select <- bugsigdbr::getSignatures(dat, tax.level = "genus", exact.tax.level=TRUE)
dat.select <- dat.select[grep("UP", names(dat.select))] #only "UP" signatures
frequencySigs(sigs=dat.select, n = 10)
}
