% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{frequencySigs}
\alias{frequencySigs}
\title{Identify the most frequently recurring taxa in bugsigdb}
\usage{
frequencySigs(sigs, n = 10)
}
\arguments{
\item{n}{Number of top taxa to return}

\item{dat}{data.frame produced by \link[bugsigdbr]{importBugSigDB}}

\item{direction}{One or both of c("increased", "decreased")}
}
\value{
integer vector of top most recurrent taxa
}
\description{
Identify the most frequently recurring taxa in bugsigdb
}
\examples{
dat <- bugsigdbr::importBugSigDB()
sigs <- extractBugSigs(dat, direction = "increased", tax.level = "genus")
frequencySigs(sigs)
}
