% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{getCriticalN}
\alias{getCriticalN}
\title{countBug counts the frequency of the most commonly identified bug in a simulated signature.
getCriticalN performs a Monte Carlo simulation to estimate the number of times the most frequent taxon is expected to be observed
in a list of signatures}
\usage{
getCriticalN(relevant.sigs, siglengths, alpha = 0.05, nsim = 1000)
}
\arguments{
\item{relevant.sigs}{a list of signatures that form the "background" from which taxa for simulated signatures will be drawn. 
These are used to estimate how frequently taxa occur}

\item{siglengths}{The sizes of signatures found in a set of related studies. Simulated signatures will match these in number and size.}

\item{alpha}{Probability at which a critical threshold will be calculated (default: 0.05)}

\item{nsim}{Number of simulations (default: 1000)}
}
\value{
The 1 - alpha quantile of Monte Carlo simulated values for the maximum number of times any taxon is identified.
}
\description{
countBug counts the frequency of the most commonly identified bug in a simulated signature.
getCriticalN performs a Monte Carlo simulation to estimate the number of times the most frequent taxon is expected to be observed
in a list of signatures
}
\details{
E.g. for alpha = 0.05, we expect only a 5% chance that any taxon will be identified N times or more.
}
\examples{
full.dat <- bugsigdbr::importBugSigDB()
my.dat <- full.dat[full.dat$Curator == "Mst Afroza Parvin", ]
relevant.dat <- full.dat[full.dat$`Body site` \%in\% my.dat$`Body site`, ]
relevant.sigs <- extractBugSigs(my.dat)
my.sigs.increased <- extractBugSigs(my.dat, direction = "increased")
(my.siglengths <- sapply(my.sigs.increased, length))
getCriticalN(relevant.sigs, my.siglengths)
# Compare to observed
frequencySigs(my.sigs.increased)
}
