% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{getMostFrequentTaxa}
\alias{getMostFrequentTaxa}
\title{Get the most frequently occurring taxa in a table of signatures}
\usage{
getMostFrequentTaxa(
  dat,
  n = 10,
  sig.type = c("both", "increased", "decreased"),
  direction.column = "Abundance in Group 1"
)
}
\arguments{
\item{dat}{A table such as output by \code{\link{bugsigdbr::importBugSigDB}}}

\item{n}{Number of most frequently occurring taxa to return}

\item{sig.type}{increased for increased in cases relative to controls, decreased for decreased in cases relative to controls, both for either}

\item{direction.column}{column containing direction information in dat}
}
\value{
A list of signatures, with PMIDs for list element names
}
\description{
Get the most frequently occurring taxa in a table of signatures
}
\examples{
full.dat <- bugsigdbr::importBugSigDB()
getMostFrequentTaxa(full.dat)
}
