% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{getMostFrequentTaxa}
\alias{getMostFrequentTaxa}
\title{Get the most frequently occurring taxa in a table of signatures}
\usage{
getMostFrequentTaxa(dat, n = 10, sig.type = c("BOTH", "UP", "DOWN"))
}
\arguments{
\item{dat}{A table such as output by \code{\link{bugSigDB::importBugSigDB}}}

\item{n}{Number of most frequently occurring taxa to return}

\item{sig.type}{UP for increased in cases relative to controls, DOWN for decreased in cases relative to controls, BOTH for either}
}
\value{
A list of signatures, with PMIDs for list element names
}
\description{
Get the most frequently occurring taxa in a table of signatures
}
\examples{
full.dat <- bugsigdbr::importBugSigDB()
getMostFrequentTaxa(full.dat)
}
