% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{simulateSignatures}
\alias{simulateSignatures}
\title{Simulate a list of signatures based on a universe of taxa equal in size and length to some smaller set of signatures}
\usage{
simulateSignatures(relevant.sigs, siglengths)
}
\arguments{
\item{relevant.sigs}{A list of signatures providing a relevant background for drawing taxa}

\item{siglengths}{An integer vector, the length of which provides the number of signatures to be simulated, and the 
integers of which provide the number of taxa to be simulated in each signature}

\item{tax.level}{Any subset of c("kingdom", "phylum", "class", "order", "family", "genus", "species", "strain")}
}
\value{
A list of signatures of the same number and individual lengths as found in my.dat
}
\description{
Simulate a list of signatures based on a universe of taxa equal in size and length to some smaller set of signatures
}
\examples{
full.dat <- importBugSigDB()
my.dat <- full.dat[full.dat$Curator == "Mst Afroza Parvin", ]
relevant.dat <- full.dat[full.dat$`Body site` \%in\% my.dat$`Body site`, ]
relevant.sigs <- extractBugSigs(relevant.dat, tax.level = "genus")
siglengths <- sapply(my.dat, length)
simulateSignatures(relevant.sigs, siglengths, tax.level = "genus")
}
